型に合わせる修正

+++ src/usecase/planusecase/planoutput/create_plan_output.go
@@ -4,10 +4,10 @@ type CreatePlanOutput struct {
 	ID                 string `json:"id,omitempty"`
 	UserID             string `json:"user_id,omitempty"`
 	Title              string `json:"title,omitempty"`
-	Category           int    `json:"category,omitempty"`
+	Category           uint16 `json:"category,omitempty"`
 	Content            string `json:"content,omitempty"`
-	Status             int    `json:"status,omitempty"`
-	ConsultationFormat int    `json:"consultation_format,omitempty"`
-	Price              int    `json:"price,omitempty"`
+	Status             uint16 `json:"status,omitempty"`
+	ConsultationFormat uint16 `json:"consultation_format,omitempty"`
+	Price              uint16 `json:"price,omitempty"`
 	ConsultationMethod int    `json:"consultation_method,omitempty"`
 }
